Introduce TouhyosanHelpers
